requirements.txt
albumentations==0.1.2
certifi==2018.8.24
chardet==3.0.4
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.3
decorator==4.3.0
easydict==1.8
graphviz==0.8.4
idna==2.6
imgaug==0.2.6
kiwisolver==1.0.1
matplotlib==3.0.0
mxboard==0.1.0
mxnet-cu90==1.3.0.post0
networkx==2.2
numpy==1.14.6
opencv-python==3.4.3.18
Pillow==5.3.0
pkg-resources==0.0.0
protobuf==3.6.1
pyparsing==2.2.2
python-dateutil==2.7.3
PyWavelets==1.0.1
requests==2.18.4
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
